refactor: Package each class
