+++ src/main/java/racingcar/movepolicy/DefaultMoveStrategy.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.movepolicy;
 
 import java.util.Random;
