+++ src/main/java/racingcar/movepolicy/MoveStrategy.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.movepolicy;
 
 public interface MoveStrategy {
     boolean move();
