+++ src/test/java/racingcar/car/CarTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.car;
 
 import org.junit.jupiter.api.Test;
