+++ src/test/java/racingcar/racing/RacingTest.java
@@ -1,6 +1,8 @@
-package racingcar;
+package racingcar.racing;
 
 import org.junit.jupiter.api.Test;
+import racingcar.movepolicy.MoveStrategy;
+import racingcar.car.Car;
 
 import java.util.List;
